requirements.txt
aiogram==3.4.1
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
sqlalchemy==2.0.23
alembic==1.12.1
